also check phone numbers in core API endpoints
